Gate promo 404/500's by donate app
