+++ readthedocs/donate/views.py
@@ -5,14 +5,18 @@ import logging
 from django.views.generic import TemplateView
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
-from django.shortcuts import redirect, get_object_or_404
+from django.shortcuts import redirect, get_object_or_404, render_to_response
+from django.template import RequestContext
 from django.core.cache import cache
 from django.http import Http404
 
 from vanilla import CreateView, ListView
 
+from readthedocs.donate.models import SupporterPromo
+from readthedocs.donate.utils import offer_promo
 from readthedocs.payments.mixins import StripeMixin
 from readthedocs.projects.models import Project
+from readthedocs.redirects.utils import get_redirect_response
 
 from .models import Supporter, SupporterPromo, CLICKS, VIEWS
 from .forms import SupporterForm, EthicalAdForm
@@ -146,3 +150,40 @@ def view_proxy(request, promo_id, hash):
         log.warning('Duplicate view logged. {count} total clicks tried.'.format(count=count))
         cache.incr(promo.cache_key(type=VIEWS, hash=hash))
     return redirect(promo.image)
+
+
+def add_promo_data(display_type):
+    promo_queryset = SupporterPromo.objects.filter(live=True, display_type=display_type)
+    promo_obj = promo_queryset.order_by('?').first()
+    if promo_obj:
+        promo_dict = offer_promo(promo_obj=promo_obj, project=None)
+    else:
+        promo_dict = None
+    return promo_dict
+
+
+def promo_500(request, template_name='500.html', **kwargs):
+    """A simple 500 handler so we get media"""
+    promo_dict = add_promo_data(display_type='error')
+    r = render_to_response(template_name,
+                           context_instance=RequestContext(request),
+                           context={
+                               'promo_data': promo_dict,
+                           })
+    r.status_code = 500
+    return r
+
+
+def promo_404(request, template_name='404.html', **kwargs):
+    """A simple 404 handler so we get media"""
+    promo_dict = add_promo_data(display_type='error')
+    response = get_redirect_response(request, path=request.get_full_path())
+    if response:
+        return response
+    r = render_to_response(template_name,
+                           context_instance=RequestContext(request),
+                           context={
+                               'promo_data': promo_dict,
+                           })
+    r.status_code = 404
+    return r
